+++ Packs/NozomiNetworks/CHANGELOG.md
@@ -1,4 +1,3 @@
 ## [Unreleased]
-- Added guardian-list-all-assets command
-- guardian-search
-- guardian-list-all-assets
+- Added a new integration with NozomiGuardian
+- Added a new Playbook for a rouge device incidents in an OT environment